do refactor
