+++ codecov.yml
@@ -2,4 +2,4 @@ fixes:
   - "pennylane_qchem/::qchem/pennylane_qchem/"
 
 ignore:
-  - "pennylane/plugins/tests/*"
+  - "pennylane/devices/tests/*"
