+++ pennylane/beta/devices/default_tensor_tf.py
@@ -28,8 +28,8 @@ except ImportError as e:
     raise ImportError("default.tensor.tf device requires TensorFlow>=2.0")
 
 from pennylane.variable import Variable
-from pennylane.beta.plugins.default_tensor import DefaultTensor
-from pennylane.plugins import tf_ops as ops
+from pennylane.beta.devices.default_tensor import DefaultTensor
+from pennylane.devices import tf_ops as ops
 
 # tolerance for numerical errors
 tolerance = 1e-10
