+++ pennylane/beta/devices/numpy_ops.py
