+++ pennylane/devices/default_qubit.py
@@ -12,8 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 r"""
-The default plugin is meant to be used as a template for writing PennyLane device
-plugins for new qubit-based backends.
+The default.qubit device is PennyLane's most basic qubit-based device.
 
 It implements the necessary :class:`~pennylane._device.Device` methods as well as some built-in
 :mod:`qubit operations <pennylane.ops.qubit>`, and provides a very simple pure state
