+++ pennylane/devices/tests/test_measurements.py
