+++ pennylane/devices/tests/test_properties.py
