+++ pennylane/devices/tf_ops.py
