+++ tests/beta/test_default_tensor.py
@@ -27,7 +27,7 @@ tensorflow = pytest.importorskip("tensorflow", minversion="2.0")
 
 import pennylane as qml
 from pennylane import numpy as np, QuantumFunctionError
-from pennylane.beta.plugins.numpy_ops import (
+from pennylane.beta.devices.numpy_ops import (
     CNOT,
     CSWAP,
     CZ,
